pyproject.toml
[project]
name = "seo-agent"
version = "0.1.0"
description = "SEO Incident Investigation Agent for Visable User Growth team"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Visable User Growth" },
]
license = { text = "Proprietary" }

dependencies = [
    "pandas>=2.2",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "typer>=0.12",
    "rich>=13.7",
    "jinja2>=3.1",
    "tenacity>=8.3",
    "python-dateutil>=2.9",
    # Google Search Console
    "google-api-python-client>=2.130",
    "google-auth>=2.30",
    "google-auth-oauthlib>=1.2",
    # Redshift
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    # Parsing rendered HTML for SSR checks
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-python-dateutil",
]

[project.scripts]
seo-agent = "seo_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["seo_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
